💥 Implement serde::Deserializer for an internal type only
This remove the serde::Deserializer impl for `&Value` for an internal
deserializer impl. We shouldn't have provide this impl for external use
anyway and in a following commit, we'll need a bit more state than just
the value. Hence we store an `Option<&Value` in the deserializer rather
than `Value`. This also means that now we can handle deserializing
`Option`, although nothing yet uses `None` values.